feat(op-revm): add testdata comparison utility for EVM execution output (bluealloy/revm#2525)
* Create test_utils.rs

* Create basic_test.json

* Create testdata_test.rs

* Create migration_example.rs

* Create fully_migrated_test.json

* Create halt_test.json

* Create revert_test.json

* Create migration_example.json

* Update Cargo.toml

* Update test_utils.rs

* Update Cargo.toml

* Update migration_example.rs

* Update Cargo.toml

* Update test_utils.rs

* Create README.md

* feat: restructure op-test and cleanup test folder

* self propagate feature to fix zepter

---------
