+++ tests/testdata/template_test.json
@@ -1,0 +1,14 @@
+{
+  "result": {
+    "Success": {
+      "reason": "Stop",
+      "gas_used": 1000,
+      "gas_refunded": 0,
+      "logs": [],
+      "output": {
+        "Call": "0x040506"
+      }
+    }
+  },
+  "state": {}
+}
